feat: animeReview on typescript
